requirements.txt
flask==2.0.3
python-dotenv==0.19.2
git+https://github.com/davidhao3300/ynab-python.git@0.0.3
pandas<1.2.0
Werkzeug==2.0.3
requests==2.23.0
git+https://github.com/python-pillow/Pillow.git@7.1.1
pdfplumber